Correction: Output
Output was previously defined as output from current machine. This is
wrong, it represents the output from the prior run of the machine.
Equations on X were also updated to reflect this. Additionally, active
words in memory was incorrectly defined as its members(i.e. code) and
not its total size.